Refactor and enhance security of the application.
This commit introduces two major improvements:

1.  **Code Refactoring:**
    - The monolithic `app.py` has been broken down into a modular structure.
    - Business logic is now separated into specialized modules (`auth`, `data_manager`, `shift_management`, `notifications`) located in a new `modules` directory.
    - Configuration variables and paths have been centralized into a `config.py` file.
    - A `requirements.txt` file has been added to manage project dependencies.

2.  **Security Enhancement:**
    - Replaced the insecure plaintext password storage with a secure hashing mechanism using `bcrypt`.
    - Added a one-time script (`migrate_passwords.py`) to migrate existing passwords to the new hashed format.
    - Replaced the insecure URL-based authentication with a standard username and password login form.
    - The authentication logic has been updated to verify credentials against the stored hashes.

These changes significantly improve the maintainability, organization, and security of the application.
